# Reduce client protocol command-reply allocs