--- a/.gitignore
+++ b/.gitignore
@@ -6,4 +6,4 @@ vendor/
 *.xlog
 *.snap.inprogress
 devdata/
-.orig
+*.orig