Mostly re-organizing code to avoid several heap allocations due to function closures. Also returning TransportWriteHandler to Node's scope since otherwise connect response is lost (i.e. reverting changes in `new_protocol` branch).

Before (`new_protocol` branch):

```
BenchmarkWebsocketHandlerCommandReply/JSON-12 	   32911	     35620 ns/op	    1803 B/op	      18 allocs/op
BenchmarkWebsocketHandlerCommandReply/Protobuf-12         	   34539	     34246 ns/op	    1787 B/op	      19 allocs/op
BenchmarkWebsocketHandlerCommandReplyV2/JSON-12           	   34080	     35402 ns/op	    1762 B/op	      17 allocs/op
BenchmarkWebsocketHandlerCommandReplyV2/Protobuf-12       	   35166	     33852 ns/op	    1762 B/op	      18 allocs/op
```

After:

```
BenchmarkWebsocketHandlerCommandReply/JSON-12 	   35349	     35350 ns/op	    1747 B/op	      15 allocs/op
BenchmarkWebsocketHandlerCommandReply/Protobuf-12         	   33186	     34638 ns/op	    1731 B/op	      16 allocs/op
BenchmarkWebsocketHandlerCommandReplyV2/JSON-12           	   35467	     34633 ns/op	    1700 B/op	      13 allocs/op
BenchmarkWebsocketHandlerCommandReplyV2/Protobuf-12       	   35568	     33382 ns/op	    1706 B/op	      14 allocs/op
```